@@ -15,31 +15,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~10.0.0",
-    "@angular/common": "~10.0.0",
-    "@angular/compiler": "~10.0.0",
-    "@angular/core": "~10.0.0",
+    "@angular/animations": "~10.1.3",
+    "@angular/common": "~10.1.3",
+    "@angular/compiler": "~10.1.3",
+    "@angular/core": "~10.1.3",
     "@angular/fire": "../dist/packages-dist",
-    "@angular/forms": "~10.0.0",
-    "@angular/platform-browser": "~10.0.0",
-    "@angular/platform-browser-dynamic": "~10.0.0",
-    "@angular/platform-server": "~10.0.0",
-    "@angular/router": "~10.0.0",
-    "@angular/service-worker": "^10.0.0",
-    "@nguniversal/express-engine": "~9.1.0",
-    "firebase": "^7.13.2",
+    "@angular/forms": "~10.1.3",
+    "@angular/platform-browser": "~10.1.3",
+    "@angular/platform-browser-dynamic": "~10.1.3",
+    "@angular/platform-server": "~10.1.3",
+    "@angular/router": "~10.1.3",
+    "@angular/service-worker": "^10.1.3",
+    "@nguniversal/express-engine": "~10.1.0",
+    "firebase": "^7.21.1",
     "proxy-polyfill": "^0.3.1",
     "rxjs": "~6.5.4",
     "tslib": "^2.0.0",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.901.0",
-    "@angular-devkit/build-angular": "~0.1000.0",
-    "@angular/cli": "~10.0.0",
-    "@angular/compiler-cli": "~10.0.0",
-    "@angular/language-service": "~10.0.0",
-    "@nguniversal/builders": "^9.1.0",
+    "@angular-devkit/architect": "~0.1001.3",
+    "@angular-devkit/build-angular": "~0.1001.3",
+    "@angular/cli": "~10.1.3",
+    "@angular/compiler-cli": "~10.1.3",
+    "@angular/language-service": "~10.1.3",
+    "@nguniversal/builders": "^10.1.0",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
